+++ src/model.jl
@@ -7,128 +7,8 @@
 include("constitutive.jl")
 include("interpolation.jl")
 include("ics_bcs.jl")
-
+include("opinf/opinf_ics_bcs.jl")
 using Base.Threads: @threads, threadid, nthreads
-using NPZ
-using PyCall
-
-function LinearOpInfRom(params::Parameters)
-    params["mesh smoothing"] = false
-    fom_model = SolidMechanics(params)
-    reference = fom_model.reference
-    opinf_model_file = params["model"]["model-file"]
-    opinf_model = NPZ.npzread(opinf_model_file)
-    basis = opinf_model["basis"]
-    _, _, reduced_dim = size(basis)
-    num_dofs = reduced_dim
-    time = 0.0
-    failed = false
-    null_vec = zeros(num_dofs)
-
-    reduced_state = zeros(num_dofs)
-    reduced_velocity = zeros(num_dofs)
-    reduced_boundary_forcing = zeros(num_dofs)
-    free_dofs = trues(num_dofs)
-    boundary_conditions = Vector{BoundaryCondition}()
-    return LinearOpInfRom(
-        opinf_model,
-        basis,
-        reduced_state,
-        reduced_velocity,
-        reduced_boundary_forcing,
-        null_vec,
-        free_dofs,
-        boundary_conditions,
-        time,
-        failed,
-        fom_model,
-        reference,
-        false,
-    )
-end
-
-function QuadraticOpInfRom(params::Parameters)
-    params["mesh smoothing"] = false
-    fom_model = SolidMechanics(params)
-    reference = fom_model.reference
-    opinf_model_file = params["model"]["model-file"]
-    opinf_model = NPZ.npzread(opinf_model_file)
-    basis = opinf_model["basis"]
-    _, _, reduced_dim = size(basis)
-    num_dofs = reduced_dim
-    time = 0.0
-    failed = false
-    null_vec = zeros(num_dofs)
-
-    reduced_state = zeros(num_dofs)
-    reduced_velocity = zeros(num_dofs)
-    reduced_boundary_forcing = zeros(num_dofs)
-    free_dofs = trues(num_dofs)
-    boundary_conditions = Vector{BoundaryCondition}()
-    return QuadraticOpInfRom(
-        opinf_model,
-        basis,
-        reduced_state,
-        reduced_velocity,
-        reduced_boundary_forcing,
-        null_vec,
-        free_dofs,
-        boundary_conditions,
-        time,
-        failed,
-        fom_model,
-        reference,
-        false,
-    )
-end
-
-function NeuralNetworkOpInfRom(params::Dict{String,Any})
-    params["mesh smoothing"] = false
-    fom_model = SolidMechanics(params)
-    reference = fom_model.reference
-    opinf_model_directory = params["model"]["model-directory"]
-    basis_file = opinf_model_directory * "/nn-opinf-basis.npz"
-    basis = NPZ.npzread(basis_file)
-    basis = basis["basis"]
-    py""" 
-    import torch
-    def get_model(model_file):
-      return torch.load(model_file)
-    """
-    ensemble_size = params["model"]["ensemble-size"]
-    model = []
-    for i in 1:ensemble_size
-      tmp =  py"get_model"(opinf_model_directory * "/stiffness-" * string(i-1) * ".pt")
-      push!(model,tmp)
-    end
-    num_dofs_per_node,num_nodes_basis,reduced_dim = size(basis)
-    num_dofs = reduced_dim
-
-    time = 0.0
-    failed = false
-    null_vec = zeros(num_dofs)
-
-    reduced_state = zeros(num_dofs)
-    reduced_velocity = zeros(num_dofs)
-    reduced_boundary_forcing = zeros(num_dofs)
-    free_dofs = trues(num_dofs)
-    boundary_conditions = Vector{BoundaryCondition}()
-    NeuralNetworkOpInfRom(
-        model,
-        basis,
-        reduced_state,
-        reduced_velocity,
-        reduced_boundary_forcing,
-        null_vec,
-        free_dofs,
-        boundary_conditions,
-        time,
-        failed,
-        fom_model,
-        reference,
-        false 
-    )
-end
 
 
 function SolidMechanics(params::Parameters)
