+++ src/model_types.jl
@@ -5,8 +5,6 @@
 # top-level Norma.jl directory.
 
 abstract type Model end
-abstract type RomModel <: Model end
-abstract type OpInfModel <: RomModel end
 using SparseArrays
 
 @enum Kinematics begin
@@ -95,70 +93,4 @@ mutable struct SolidMechanics <: Model
     kinematics::Kinematics
 end
 
-mutable struct QuadraticOpInfRom <: OpInfModel
-    opinf_rom::Dict{Any,Any}
-    basis::Array{Float64}
-    reduced_state::Vector{Float64}
-    reduced_velocity::Vector{Float64}
-    reduced_boundary_forcing::Vector{Float64}
-    #internal_force not used, but include to ease interfacing in Schwarz
-    internal_force::Vector{Float64}
-    free_dofs::BitVector
-    boundary_conditions::Vector{BoundaryCondition}
-    time::Float64
-    failed::Bool
-    fom_model::SolidMechanics
-    reference::Matrix{Float64}
-    inclined_support::Bool
-end
 
-mutable struct GalerkinRom <: OpInfModel
-    opinf_rom::Dict{Any,Any}
-    basis::Array{Float64}
-    reduced_state::Vector{Float64}
-    reduced_velocity::Vector{Float64}
-    reduced_boundary_forcing::Vector{Float64}
-    #internal_force not used, but include to ease interfacing in Schwarz
-    internal_force::Vector{Float64}
-    free_dofs::BitVector
-    boundary_conditions::Vector{BoundaryCondition}
-    time::Float64
-    failed::Bool
-    fom_model::SolidMechanics
-    reference::Matrix{Float64}
-    inclined_support::Bool
-end
-
-mutable struct LinearOpInfRom <: OpInfModel
-    opinf_rom::Dict{Any,Any}
-    basis::Array{Float64}
-    reduced_state::Vector{Float64}
-    reduced_velocity::Vector{Float64}
-    reduced_boundary_forcing::Vector{Float64}
-    #internal_force not used, but include to ease interfacing in Schwarz
-    internal_force::Vector{Float64}
-    free_dofs::BitVector
-    boundary_conditions::Vector{BoundaryCondition}
-    time::Float64
-    failed::Bool
-    fom_model::SolidMechanics
-    reference::Matrix{Float64}
-    inclined_support::Bool
-end
-
-mutable struct NeuralNetworkOpInfRom <: OpInfModel
-    nn_model::Any
-    basis::Array{Float64}
-    reduced_state::Vector{Float64}
-    reduced_velocity::Vector{Float64}
-    reduced_boundary_forcing::Vector{Float64}
-    #internal_force not used, but include to ease interfacing in Schwarz
-    internal_force::Vector{Float64}
-    free_dofs::BitVector
-    boundary_conditions::Vector{BoundaryCondition}
-    time::Float64
-    failed::Bool
-    fom_model::SolidMechanics
-    reference::Matrix{Float64}
-    inclined_support::Bool
-end
